fix base-class-of-parameter
