+++ lib/gmo-payment-carrier.rb
@@ -5,6 +5,7 @@ require "faraday"
 require "gmo-payment-carrier/version"
 require "gmo-payment-carrier/const"
 require "gmo-payment-carrier/client"
+require "gmo-payment-carrier/parameter"
 require "gmo-payment-carrier/converter"
 require "gmo-payment-carrier/au"
 require "gmo-payment-carrier/docomo"
