+++ spec/gmo-payment-carrier/au/parameter_spec.rb
@@ -1,13 +1,14 @@
 require "spec_helper"
 
 describe GMOPaymentCarrier::AU::Parameter do
-  let(:target) { GMOPaymentCarrier::AU::Parameter.new }
+  describe 'entry_tran_au_continuance' do
+    let(:target) { GMOPaymentCarrier::AU::Parameter.new(api_kind: :entry_tran_au_continuance) }
 
-  it do
-    target.api_kind = :entry_tran_au_continuance
-    target.shop_id = nil
+    it do
+      target.shop_id = nil
 
-    target.valid?
-    expect(target.errors[:shop_id].present?).to eq(true)
+      target.valid?
+      expect(target.errors[:shop_id].present?).to eq(true)
+    end
   end
 end
